+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.5.0] - 2018-10-19
+
+### Added
+- DatabaseHelper added from operations package.
+- SpamHelper::is_spam(). Checks all fields for spam score now.
+- SpamHelper::has_url().
+
+### Changed
+- SpamHelper revamp.
+
 ## [1.4.0] - 2018-10-09
 
 ### Added
@@ -111,7 +121,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.1] - 2018-01-17
 
-[Unreleased]: https://github.com/clsimplex/tuxedo-helpers/compare/1.4.0...develop
+[Unreleased]: https://github.com/clsimplex/tuxedo-helpers/compare/1.5.0...develop
+[1.5.0]: https://github.com/clsimplex/tuxedo-helpers/compare/1.4.0...1.5.0
 [1.4.0]: https://github.com/clsimplex/tuxedo-helpers/compare/1.3.2...1.4.0
 [1.3.2]: https://github.com/clsimplex/tuxedo-helpers/compare/1.3.1...1.3.2
 [1.3.1]: https://github.com/clsimplex/tuxedo-helpers/compare/1.3.0...1.3.1
@@ -124,4 +135,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.0.3]: https://github.com/clsimplex/tuxedo-helpers/compare/0.0.2...0.0.3
 [0.0.2]: https://github.com/clsimplex/tuxedo-helpers/compare/0.0.1...0.0.2
 [0.0.1]: https://github.com/clsimplex/tuxedo-helpers/releases/0.0.1
-